+++ R/lasclip.r
@@ -104,7 +104,7 @@
 #'
 #' # Extract the polygons, write them in files name after the lake names, do not load anything in R
 #' output_files(ctg) <- paste0(tempfile(), "_{LAKENAME_1}")
-#' lasclip(ctg, lake)
+#' lasclip(ctg, lakes)
 #'
 #' \dontrun{
 #' plot(subset1)