+++ man-roxygen/param-treetops.R
@@ -1,0 +1,2 @@
+#' @param treetops \code{SpatialPolygonsDataFrame}. Can be computed with
+#' \link[lidR:tree_detection]{tree_detection} or read from an external shapefile.